simplified call_api for common use cases
The *.create and *.update family of API calls expect a single BaseObject
instance as the request's sole child. This used to require a call like:

  freshbooks.call_api('time_entry.create', {'dummy': time_entry_object})

Where the 'dummy' dictionary key is not used.
This common use case is now simplified to:

  freshbooks.call_api('time_entry.create', time_entry_object)

The old call style is no longer supported.
This doesn't break backward-compatibility, because passing a BaseObject
was previously throwing an exception (see previous commit).