add update example to doc
Documented the new call_api call style (see previous commit) through an
example describing how to update a client.

This also helps with an issue I first had with the library: a customary
look at the documentation and code made me think that the library only
offered a read-only access to the FreshBooks data.